+++ packages/ast-spec/src/element/AccessorProperty/fixtures/key-private/fixture.ts
@@ -1,0 +1,3 @@
+class Foo {
+  accessor #foo = 2;
+}