+++ packages/typescript-estree/src/parseSettings/warnAboutTSVersion.ts
@@ -3,8 +3,8 @@ import * as ts from 'typescript';
 
 import type { ParseSettings } from './index';
 /**
- * This needs to be kept in sync with the top-level README.md in the
- * typescript-eslint monorepo
+ * This needs to be kept in sync with /docs/maintenance/VERSIONING.md
+ * in the typescript-eslint monorepo
  */
 const SUPPORTED_TYPESCRIPT_VERSIONS = '>=3.3.1 <5.0.0';